requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
aiofiles=0.8.0=pypi_0
anyio=3.5.0=pypi_0
argcomplete=2.0.0=pypi_0
boto3=1.21.40=pypi_0
botocore=1.24.40=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.3.29=h06a4308_0
certifi=2020.6.20=pyhd3eb1b0_3
charset-normalizer=2.0.12=pypi_0
cycler=0.11.0=pypi_0
deprecated=1.2.13=pypi_0
dill=0.3.4=pypi_0
dnspython=2.2.1=pypi_0
eventlet=0.33.0=pypi_0
fiftyone=0.15.0.1=pypi_0
fiftyone-brain=0.8.1=pypi_0
fiftyone-db=0.3.0=pypi_0
fonttools=4.32.0=pypi_0
future=0.18.2=pypi_0
glob2=0.7=pypi_0
greenlet=1.1.2=pypi_0
h11=0.12.0=pypi_0
httpcore=0.14.7=pypi_0
httpx=0.22.0=pypi_0
idna=3.3=pypi_0
imageio=2.16.2=pypi_0
jinja2=3.1.1=pypi_0
jmespath=1.0.0=pypi_0
joblib=1.1.0=pypi_0
kaleido=0.2.1=pypi_0
kiwisolver=1.4.2=pypi_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libstdcxx-ng=9.3.0=hd4cf53a_17
libuuid=1.0.3=h7f8727e_2
markupsafe=2.1.1=pypi_0
matplotlib=3.5.1=pypi_0
mongoengine=0.20.0=pypi_0
motor=2.5.1=pypi_0
ncurses=6.3=h7f8727e_2
ndjson=0.3.1=pypi_0
networkx=2.8=pypi_0
numpy=1.22.3=pypi_0
opencv-python-headless=4.5.5.64=pypi_0
openssl=1.1.1n=h7f8727e_0
packaging=21.3=pypi_0
pandas=1.4.2=pypi_0
patool=1.12=pypi_0
pillow=9.1.0=pypi_0
pip=21.2.4=pypi_0
plotly=4.14.3=pypi_0
pprintpp=0.4.0=pypi_0
psutil=5.9.0=pypi_0
pymongo=3.12.3=pypi_0
pyparsing=3.0.8=pypi_0
python=3.10.4=h12debd9_0
python-dateutil=2.8.2=pypi_0
pytz=2022.1=pypi_0
pytz-deprecation-shim=0.1.0.post0=pypi_0
pywavelets=1.3.0=pypi_0
pyyaml=6.0=pypi_0
readline=8.1.2=h7f8727e_1
requests=2.27.1=pypi_0
retrying=1.3.3=pypi_0
rfc3986=1.5.0=pypi_0
s3transfer=0.5.2=pypi_0
scikit-image=0.19.2=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.8.0=pypi_0
setuptools=61.2.0=pypi_0
six=1.16.0=pypi_0
sniffio=1.2.0=pypi_0
sortedcontainers=2.4.0=pypi_0
sqlite=3.38.2=hc218d9a_0
tabulate=0.8.9=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2022.4.8=pypi_0
tk=8.6.11=h1ccaba5_0
tornado=6.1=pypi_0
tzdata=2022.1=pypi_0
tzlocal=4.2=pypi_0
universal-analytics-python3=1.1.1=pypi_0
urllib3=1.26.9=pypi_0
voxel51-eta=0.6.5=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wrapt=1.14.0=pypi_0
xmltodict=0.12.0=pypi_0
xz=5.2.5=h7b6447c_0
zlib=1.2.12=h7f8727e_0
